+++ demo/rgl.r
@@ -1,8 +1,18 @@
 # all rgl demos
 demo(hist3d)
 demo(abundance)
-demo(regression)
 demo(lsystem)
 demo(subdivision)
-# requires MASS library
+# requires MASS package
 demo(bivar)
+demo(envmap)
+demo(flag)
+demo(lollipop3d)
+demo(mouseCallbacks)
+demo(shapes3d)
+demo(shinyDemo)
+demo(shinyMouse)
+demo(shinyTabs)
+demo(shinyToggle)
+demo(simpleShinyRgl)
+demo(stereo)